insightease-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
email-validator
bcrypt
python-multipart
pandas
numpy
scikit-learn
openpyxl
pymysql
cryptography
aiomysql
greenlet
python-jose[cryptography]
passlib[bcrypt]
aiofiles
openai
prophet
matplotlib
seaborn
reportlab
alembic
